feat(seedgo): checker consolidation: testing→error_handling, test_coverage merged into test_quality v4.0
