+++ src/aipass/seedgo/apps/handlers/aipass_standards/error_handling_content.py
@@ -2,164 +2,92 @@
 # Name: error_handling_content.py
 # Description: Error Handling Standards Content Handler
 # Version: 1.0.0
-# Created: 2026-03-05
-# Modified: 2026-03-05
+# Created: 2026-03-27
+# Modified: 2026-03-27
 # =============================================
 
 """
 Error Handling Standards Content Handler
 
-Provides formatted error handling standards content (3-tier architecture).
+Provides formatted error handling standards content.
 Module orchestrates, handler implements.
 """
 
 from aipass.seedgo.apps.handlers.json import json_handler
 
+
 def get_error_handling_standards() -> str:
-    """Return formatted error handling standards content with Rich markup
+    """Return formatted error_handling standards content with Rich markup
 
     Returns:
         str: Formatted standards text with Rich styling
     """
     lines = [
-        "[bold red]3-TIER LOGGING ARCHITECTURE[/bold red]",
+        "[bold cyan]CORE PRINCIPLE:[/bold cyan]",
+        "  Errors must tell the truth. Silent failures hide bugs and erode",
+        "  trust in the codebase. Every except block must handle the error",
+        "  meaningfully — log it, return it, or re-raise it.",
         "",
-        "[yellow]CORE PRINCIPLE:[/yellow] Logging responsibility follows architectural hierarchy",
+        "[bold cyan]WHAT IT CHECKS:[/bold cyan]",
+        "  File-level analysis of try/except patterns:",
         "",
-        "[bold cyan]TIER 1: ENTRY POINTS[/bold cyan] (flow.py, seedgo.py, etc.)",
-        "  [dim]Prax Import:[/dim] YES (minimal)",
-        "  [dim]Logging Scope:[/dim] Operational only (discovery, routing, help)",
-        "  [red]✗ NO business error logging[/red]",
+        "  [yellow]Detection:[/yellow]",
+        "  Scans all Python files for [dim]try/except[/dim] blocks",
+        "  Flags silent failures: [dim]except: pass[/dim] or [dim]except Exception: pass[/dim]",
+        "  Skips files with no try/except (not applicable)",
         "",
-        "[bold cyan]TIER 2: MODULES[/bold cyan] (apps/modules/*.py)",
-        "  [dim]Prax Import:[/dim] [green]YES (REQUIRED)[/green]",
-        "  [dim]Logging Scope:[/dim] [green]ALL BUSINESS LOGGING[/green]",
-        "  [green]✓ Log ALL errors: logger.error(f'[{MODULE_NAME}] {msg}')[/green]",
-        "  [green]✓ Catch exceptions from handlers[/green]",
-        "  [green]✓ Provide context for debugging[/green]",
-        "  [green]✓ Return bool to entry point[/green]",
+        "[bold cyan]VIOLATIONS:[/bold cyan]",
         "",
-        "[bold cyan]TIER 3: HANDLERS[/bold cyan] (apps/handlers/**/*.py)",
-        "  [dim]Prax Import:[/dim] [green]ALLOWED[/green]",
-        "  [dim]Logging Scope:[/dim] [green]info, warning, error[/green]",
-        "  [green]✓ logger.info() — operational visibility[/green]",
-        "  [green]✓ logger.warning() — non-critical issues[/green]",
-        "  [green]✓ logger.error() — error context at point of failure[/green]",
-        "  [green]✓ Return status dicts or raise exceptions[/green]",
-        "  [green]✓ Testable in isolation[/green]",
+        "  [red]Fail:[/red] Silent failure detected — [dim]except: pass[/dim] with no",
+        "  logging, return, or re-raise",
         "",
-        "─" * 70,
+        "  [green]Pass:[/green] All except blocks have meaningful handling",
+        "  (logging, return values, re-raise, or other statements)",
         "",
-        "[bold yellow]HANDLER RETURN PATTERNS:[/bold yellow]",
+        "[bold cyan]CODE EXAMPLES:[/bold cyan]",
         "",
-        "[dim]Option 1: Status Dict[/dim]",
-        "  return {'success': True, 'data': result, 'error': None}",
-        "  return {'success': False, 'data': None, 'error': 'Error message'}",
-        "",
-        "[dim]Option 2: Raise Exceptions[/dim]",
-        "  raise ValueError('Invalid input')",
-        "  raise FileNotFoundError('Config missing')",
-        "",
-        "─" * 70,
-        "",
-        "[bold yellow]MODULE ERROR HANDLING PATTERN:[/bold yellow]",
-        "",
-        "  [dim]MODULE_NAME = \"module_name\"[/dim]",
-        "",
+        "  [green]Good:[/green]",
         "  [dim]try:[/dim]",
-        "    [dim]result = handler_function(**args)[/dim]",
-        "    [dim]if result['success']:[/dim]",
-        "      [dim]logger.info(f\"[{MODULE_NAME}] Success\")[/dim]",
-        "    [dim]else:[/dim]",
-        "      [dim]logger.error(f\"[{MODULE_NAME}] {result['error']}\")[/dim]",
+        "    [dim]result = api_call()[/dim]",
         "  [dim]except Exception as e:[/dim]",
-        "    [dim]logger.error(f\"[{MODULE_NAME}] Error: {e}\")[/dim]",
+        "    [dim]logger.error(f'API call failed: {e}')[/dim]",
+        "    [dim]return {'success': False, 'error': str(e)}[/dim]",
         "",
-        "─" * 70,
+        "  [red]Bad:[/red]",
+        "  [dim]try:[/dim]",
+        "    [dim]result = api_call()[/dim]",
+        "  [dim]except:[/dim]",
+        "    [dim]pass  # Silent failure — error swallowed[/dim]",
         "",
-        "[bold yellow]LOG LEVEL GUIDELINES: ERROR vs WARNING[/bold yellow]",
+        "[bold cyan]PHILOSOPHY:[/bold cyan]",
         "",
-        "[dim]Core Distinction:[/dim] ERROR triggers Prax escalation, WARNING does not",
+        "  [yellow]Fix error handling BEFORE fixing bugs.[/yellow]",
+        "  If errors lie, you can't debug effectively.",
+        "  Honest errors → faster debugging → reliable code.",
         "",
-        "[red]logger.error()[/red] → [bold]System failures[/bold]",
-        "  File I/O errors, crashes, dependency failures, unexpected exceptions",
+        "  The debug cycle:",
+        "  [dim]1. Fix error handling — make errors tell the truth[/dim]",
+        "  [dim]2. Then fix the actual bug[/dim]",
+        "  [dim]3. See clean pass with honest outputs[/dim]",
         "",
-        "[yellow]logger.warning()[/yellow] → [bold]User input issues[/bold]",
-        "  'Plan not found', 'Field required', 'Invalid format', 'Already exists'",
+        "[bold cyan]SCORING:[/bold cyan]",
         "",
-        "[dim]Pattern:[/dim]",
-        "  [red]# System failure → ERROR (Prax escalates)[/red]",
-        "  [dim]logger.error(f\"[{MODULE_NAME}] Failed to write file: {e}\")[/dim]",
+        "  Score = (passed_checks / total_checks) * 100",
+        "  Overall pass threshold: [yellow]75%[/yellow]",
+        "  No try/except blocks: [green]100%[/green] (not applicable)",
         "",
-        "  [yellow]# User input issue → WARNING (no escalation)[/yellow]",
-        "  [dim]logger.warning(f\"[{MODULE_NAME}] Plan {plan_id} not found\")[/dim]",
+        "[yellow]SCOPE:[/yellow]",
+        "  AUDIT_SCOPE = [bold]all_files[/bold]",
+        "  Runs on every Python file. Entry point: [dim]check_module()[/dim]",
         "",
-        "[green]✓ CLI can still show 'ERROR' text to user (feedback)[/green]",
-        "[green]✓ System logs use WARNING level (no Prax escalation)[/green]",
-        "",
-        "─" * 70,
-        "",
-        "[bold yellow]WHY THIS PATTERN:[/bold yellow]",
-        "",
-        "[green]✓ Scalability:[/green] 17 branches × 50+ handlers = 850 files to manage",
-        "[green]✓ Auditability:[/green] Check modules/ only - all logging in one place",
-        "[green]✓ Testability:[/green] Automated scans verify compliance",
-        "[green]✓ Consistency:[/green] Single pattern across entire ecosystem",
-        "",
-        "─" * 70,
-        "",
-        "[bold yellow]VALIDATION RULES:[/bold yellow]",
-        "",
-        "[dim]# All modules MUST import Prax[/dim]",
-        "[dim]grep -r \"from aipass.prax\" apps/modules/*.py[/dim]",
-        "",
-        "[dim]# Handlers MAY import Prax and use all log levels[/dim]",
-        "[dim]grep -r \"from aipass.prax\" apps/handlers/**/*.py  # Allowed[/dim]",
-        "",
-        "[dim]# Handlers MUST NOT use stdlib logging.getLogger()[/dim]",
-        "[dim]grep -r \"logging.getLogger\" apps/handlers/**/*.py  # Should find NOTHING[/dim]",
-        "",
-        "─" * 70,
+        "[bold cyan]BYPASS:[/bold cyan]",
+        "  Via [dim].seedgo/bypass.json[/dim] — supports standard-level,",
+        "  file-level, and line-level bypass rules",
         "",
         "[bold cyan]REFERENCE:[/bold cyan]",
-        "  [dim]See: seedgo standards pack (error_handling)[/dim]",
-        "  [dim]See: src/aipass/flow/apps/modules/ (working example)[/dim]",
-        "  [dim]See: src/aipass/seedgo/apps/standards/aipass/modules/ (reference implementation)[/dim]",
-        "",
-        "[bold]Decision:[/bold] 3-Tier approved 2025-11-21, ERROR/WARNING 2026-01-31",
-        "[bold]Status:[/bold] Active - All 18 branches",
-        "",
-        "─" * 70,
-        "",
-        "[bold magenta]SERVICE LOGGING STANDARDS[/bold magenta]",
-        "",
-        "[yellow]For services with user-facing interactions (bridges, APIs, bots):[/yellow]",
-        "",
-        "[bold cyan]1. LOG MEANINGFUL CONTENT[/bold cyan]",
-        "  [green]✓ 'Received from user X: hello' + 'Sent: Hi!'[/green]",
-        "  [red]✗ 'Message received' + 'Response sent'[/red]",
-        "",
-        "[bold cyan]2. AUDIT TRAILS REQUIRED[/bold cyan]",
-        "  • Chat messages and responses",
-        "  • API requests and responses",
-        "  • User interactions with context",
-        "",
-        "[bold cyan]3. LOG LOCATIONS[/bold cyan]",
-        "  [dim]~/system_logs/[/dim]         → Prax operational logs",
-        "  [dim]<branch>/logs/[/dim]         → Service content logs",
-        "  [dim]<branch>/logs/audit/[/dim]   → Interaction audit trails",
-        "",
-        "[bold cyan]4. CONTENT LOG FORMAT[/bold cyan]",
-        "  [dim]{[/dim]",
-        "    [dim]'timestamp': '2026-02-04T10:30:00Z',[/dim]",
-        "    [dim]'direction': 'inbound',[/dim]",
-        "    [dim]'user': 'user_id',[/dim]",
-        "    [dim]'content': 'message text',[/dim]",
-        "    [dim]'response': 'bot response',[/dim]",
-        "    [dim]'metadata': {...}[/dim]",
-        "  [dim]}[/dim]",
-        "",
-        "[bold]Added:[/bold] 2026-02-04 - Service logging for user-facing interactions",
+        "  [dim]Checker: error_handling_check.py[/dim]",
+        "  [dim]Standard label: ERROR_HANDLING[/dim]",
+        "  [dim]Previously: testing_check.py (renamed for clarity)[/dim]",
     ]
 
     json_handler.log_operation("standard_content_queried", {"standard": "error_handling"})
